Refactor contract error handling
- Refactor contract error handling to its own file for better readability. Some re-structuring of the ``raise_contract_logic_error_on_revert()`` method could probably also be done but for now just move this over to its own spot in the ``_utils``.
