+++ web3/_utils/contract_error_handling.py
@@ -1,0 +1,108 @@
+from eth_abi import (
+    abi,
+)
+from eth_utils import (
+    to_bytes,
+)
+
+from web3.exceptions import (
+    ContractCustomError,
+    ContractLogicError,
+    OffchainLookup,
+)
+from web3.types import (
+    RPCResponse,
+)
+
+# func selector for "Error(string)"
+SOLIDITY_ERROR_FUNC_SELECTOR = "0x08c379a0"
+
+# --- CCIP Read - EIP-3668 --- #
+# the first 4 bytes of keccak hash (func selector) for:
+# "OffchainLookup(address,string[],bytes,bytes4,bytes)"
+OFFCHAIN_LOOKUP_FUNC_SELECTOR = "0x556f1830"
+OFFCHAIN_LOOKUP_FIELDS = {
+    "sender": "address",
+    "urls": "string[]",
+    "callData": "bytes",
+    "callbackFunction": "bytes4",
+    "extraData": "bytes",
+}
+
+
+def raise_contract_logic_error_on_revert(response: RPCResponse) -> RPCResponse:
+    """
+    Reverts contain a `data` attribute with the following layout:
+        "Reverted "
+        Function selector for Error(string): 08c379a (4 bytes)
+        Data offset: 32 (32 bytes)
+        String length (32 bytes)
+        Reason string (padded, use string length from above to get meaningful part)
+
+    See also https://solidity.readthedocs.io/en/v0.6.3/control-structures.html#revert
+    """
+    if not isinstance(response["error"], dict):
+        raise ValueError("Error expected to be a dict")
+
+    data = response["error"].get("data", "")
+
+    # Ganache case:
+    if isinstance(data, dict) and response["error"].get("message"):
+        raise ContractLogicError(
+            f'execution reverted: {response["error"]["message"]}', data=data
+        )
+
+    # Parity/OpenEthereum case:
+    if data.startswith("Reverted "):
+        # "Reverted", function selector and offset are always the same for revert errors
+        prefix = f"Reverted {SOLIDITY_ERROR_FUNC_SELECTOR}"
+        data_offset = ("00" * 31) + "20"  # 0x0000...0020 (32 bytes)
+        revert_pattern = prefix + data_offset
+
+        if data.startswith(revert_pattern):
+            # if common revert pattern
+            string_length = int(
+                data[len(revert_pattern) : len(revert_pattern) + 64], 16
+            )
+            reason_as_hex = data[
+                len(revert_pattern) + 64 : len(revert_pattern) + 64 + string_length * 2
+            ]
+        elif data.startswith("Reverted 0x"):
+            # Special case for this form: 'Reverted 0x...'
+            reason_as_hex = data.split(" ")[1][2:]
+        else:
+            raise ContractLogicError("execution reverted", data=data)
+
+        reason_string = bytes.fromhex(reason_as_hex).decode("utf8")
+        raise ContractLogicError(f"execution reverted: {reason_string}", data=data)
+
+    # --- EIP-3668 | CCIP Read --- #
+    if data[:10] == OFFCHAIN_LOOKUP_FUNC_SELECTOR:
+        parsed_data_as_bytes = to_bytes(hexstr=data[10:])
+        abi_decoded_data = abi.decode(
+            list(OFFCHAIN_LOOKUP_FIELDS.values()), parsed_data_as_bytes
+        )
+        offchain_lookup_payload = dict(
+            zip(OFFCHAIN_LOOKUP_FIELDS.keys(), abi_decoded_data)
+        )
+        raise OffchainLookup(offchain_lookup_payload, data=data)
+
+    # Solidity 0.8.4 introduced custom error messages that allow args to
+    # be passed in (or not). See:
+    # https://blog.soliditylang.org/2021/04/21/custom-errors/
+    if len(data) >= 10 and not data[:10] == SOLIDITY_ERROR_FUNC_SELECTOR:
+        # Raise with data as both the message and the data for backwards
+        # compatibility and so that data can be accessed via 'data' attribute
+        # on the ContractCustomError exception
+        raise ContractCustomError(data, data=data)
+
+    # Geth case:
+    if "message" in response["error"] and response["error"].get("code", "") == 3:
+        message = response["error"]["message"]
+        raise ContractLogicError(message, data=data)
+
+    # Geth Revert without error message case:
+    if "execution reverted" in response["error"].get("message"):
+        raise ContractLogicError("execution reverted", data=data)
+
+    return response
